backend/requirements.txt
fastapi
uvicorn
pandas
google-genai
python-multipart
python-dotenv
google-cloud-storage
gcsfs